Refactor ProductList to handle item quantities
Edited the ProductList component reference
in App component, adding more props to pass to
handle the removal and addition of item quantities
whenever we add an item to the cart. ProductList
was refactored to also work with this change,
rendering a different styled button when a product
is added, as opposed to the default add to cart
button.