Style header productlist and app components
Adjusted images to be full width instead of a
custom size, and that along with adjustments to
padding fixed a wrapping issue to maintain 3 per
row. And then styling was applied to the add to
cart button and margin bottom to the header,
bringing the result closer to the finished
product.