Style components
Styled header h2 to move it closer to the images
to match the design look. Adjusted the cart
component width to more closely match the design
in how much width its taking up. And styled both
button states to match the design look. The state
of the button with items added is incomplete.