Mobile styling for Modal component
This wraps up majority styling for the project.
Refactoring was necessary in certain sections
to fix issues that arose if too many items were
added. The changes applied like overflow-y-auto
and the maxHeight calc helped in dynamically
adjusting the modal containers accordingly, so
never gets pushed off the screen. This helped
maintain the other styles and positioning.